mysite/requirements.txt
asgiref==3.5.1
boto3==1.24.2
botocore==1.27.2
Django==4.0.4
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-metadata==4.11.4
jmespath==1.0.0
Markdown==3.3.7
psycopg2==2.9.3
python-dateutil==2.8.2
python-magic==0.4.26
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
zipp==3.8.0
